better variable naming and running through the code to make sure operations are correctly applied

+++ goal_directed_generation/dataclass.py
@@ -11,7 +11,7 @@ class ReinforcementLearningParameters:
     batch_size: int
     learning_rate: float = 0.0001
     margin_threshold: int = 50
-    sigma: int = 128
+    sigma: float = 128.0
     augmented_memory: bool = True
     augmentation_rounds: int = 100
     selective_memory_purge: bool = True
